requirements.txt
asgiref==3.5.2
boto3==1.24.34
botocore==1.27.34
distlib==0.3.5
dj-database-url==1.0.0
Django==4.0.4
django-cors-headers==3.11.0
django-environ==0.9.0
django-on-heroku==1.1.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.7.1
gunicorn==20.1.0
jmespath==1.0.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
virtualenv==20.16.2
whitenoise==6.2.0
